+++ src/main/java/me/webhead1104/township/commands/CoinsCommand.java
@@ -7,13 +7,13 @@ import org.bukkit.entity.Player;
 
 public final class CoinsCommand {
 
-    public static void addCoins(Player player, long coins) {
+    public static void addCoins(Player player, int coins) {
         User user = Township.getUserManager().getUser(player.getUniqueId());
         user.setCoins(user.getCoins() + coins);
         player.sendMessage(Msg.format("<green>Added %d coins!", coins));
     }
 
-    public static void removeCoins(Player player, long coins) {
+    public static void removeCoins(Player player, int coins) {
         User user = Township.getUserManager().getUser(player.getUniqueId());
         if (user.getCoins() - coins < 0) {
             player.sendMessage(Msg.format("<red>Unable to remove %d coins!", coins));
@@ -23,7 +23,7 @@ public final class CoinsCommand {
         player.sendMessage(Msg.format("<green>Removed %d coins!", coins));
     }
 
-    public static void setCoins(Player player, long coins) {
+    public static void setCoins(Player player, int coins) {
         User user = Township.getUserManager().getUser(player.getUniqueId());
         user.setCoins(coins);
         player.sendMessage(Msg.format("<green>Set coins to %d!", coins));
@@ -31,7 +31,7 @@ public final class CoinsCommand {
 
     public static void getCoins(Player player) {
         User user = Township.getUserManager().getUser(player.getUniqueId());
-        long coins = user.getCoins();
+        int coins = user.getCoins();
         player.sendMessage(Msg.format("<green>You have %d coins!", coins));
     }
 }
